--- ui/src/utils/context/DarkModeProvider/utils/constant.ts
@@ -1,3 +1,0 @@
-const DARK_MODE_STORAGE_KEY = "darkMode";
-
-export { DARK_MODE_STORAGE_KEY };
