--- ui/src/utils/context/DarkModeProvider/utils/types.ts
@@ -1,9 +1,0 @@
-import { MapToken, SeedToken } from "antd/es/theme/interface";
-
-interface DarkModeContextType {
-	isDarkMode: boolean;
-	toggleTheme: () => void;
-	algorithm: (token: SeedToken) => MapToken;
-}
-
-export type { DarkModeContextType };
