--- ui/src/utils/context/SearchModalProvider/utils/types.ts
@@ -1,6 +1,0 @@
-interface SearchModalContextType {
-	isModalOpen: boolean;
-	handleModalOpen: () => void;
-}
-
-export type { SearchModalContextType };
